moved collections to collection folder updated test references
