+++ src/bic/collection/stars.js
