+++ tests/unit/collection-datasuitcases.js
@@ -1,7 +1,7 @@
 define(['Squire', 'backbone', 'sinon', 'chai'], function (Squire, Backbone, sinon, chai) {
   'use strict';
 
-  var CONTEXT = 'tests/unit/collection-datasuitcases.js';
+  var CONTEXT = 'tests/unit/collection/datasuitcases.js';
   var should = chai.should();
 
   describe('Collection - DataSuitcases', function () {
@@ -17,7 +17,7 @@ define(['Squire', 'backbone', 'sinon', 'chai'], function (Squire, Backbone, sino
       injector.mock('backbone', Backbone);
 
       injector.mock('bic/model-datasuitcase', Backbone.Model);
-      injector.require(['bic/collection-datasuitcases'], function (required) {
+      injector.require(['bic/collection/datasuitcases'], function (required) {
         Collection = required;
         done();
       });
