+++ tests/unit/collection-forms.js
@@ -1,7 +1,7 @@
 define(['Squire', 'backbone', 'chai'], function (Squire, Backbone, chai) {
   'use strict';
 
-  var CONTEXT = 'tests/unit/collection-forms.js';
+  var CONTEXT = 'tests/unit/collection/forms.js';
   var should = chai.should();
 
   describe('Collection - Forms', function () {
@@ -20,7 +20,7 @@ define(['Squire', 'backbone', 'chai'], function (Squire, Backbone, chai) {
       injector.mock('bic/model-form', Backbone.Model);
       injector.mock('bic/api', {});
 
-      injector.require(['bic/collection-forms'], function (rCol) {
+      injector.require(['bic/collection/forms'], function (rCol) {
         Collection = rCol;
         done();
       });
