+++ tests/unit/model-datasuitcase.js
@@ -1,7 +1,7 @@
 define(['Squire', 'backbone', 'chai'], function (Squire, Backbone, chai) {
   'use strict';
 
-  var CONTEXT = 'tests/unit/collection-forms.js';
+  var CONTEXT = 'tests/unit/collection/forms.js';
   var should = chai.should();
 
   describe('Model - DataSuitcase', function () {
